Follow the Money: The Link between Passwords and Terrorism
This co-speaker session will lead the audience on an exploration of motivations behind breaches and what actually happens to the money that nefarious actors gain by exploiting compromised credentials. It is common knowledge that compromised credentials are responsible for the majority of high impact breaches. A realm that is less known is what do the bad guys actually do with this money.

Pre-Requisites: General interest in Cyber Security


https://youtu.be/Zs0Y-6VZ2fc